+++ src/Apple/ApnPush/Connection/AbstractConnection.php
@@ -11,9 +11,6 @@
 
 namespace Apple\ApnPush\Connection;
 
-use RequestStream\Stream\Context;
-use Apple\ApnPush\Exceptions\CertificateFileNotFoundException;
-
 /**
  * Abstract core connection for Apple push notification
  */
@@ -59,33 +56,6 @@ abstract class AbstractConnection implements ConnectionInterface
         if ($sandboxMode !== null) {
             $this->setSandboxMode($sandboxMode);
         }
-    }
-
-    /**
-     * Initialize connection
-     */
-    protected function initConnection()
-    {
-        if ($this->socketConnection->is(false)) {
-            return $this;
-        }
-
-        if (!$this->certificateFile) {
-            throw new CertificateFileNotFoundException('Not found certificate file. Please set certificate file to connection.');
-        }
-
-        $context = new Context;
-        $context->setOptions('ssl', 'local_cert', $this->certificateFile);
-
-        if ($this->certificatePassPhrase) {
-            $context->setOptions('ssl', 'passphrase', $this->certificatePassPhrase);
-        }
-
-        $this->socketConnection
-            ->setTransport('ssl')
-            ->setTarget($this->getConnectionUrl())
-            ->setPort($this->getConnectionPort())
-            ->setContext($context);
     }
 
     /**
